requirements.txt
torch==2.0.1
Pillow==9.5.0
timm==0.9.2
open-clip-torch==2.20.0
torchvision==0.15.2
numpy==1.24.3
tqdm==4.65.0
omegaconf==2.3.0
pytorch_lightning==1.6.5
accelerate==0.16.0
diffusers==0.16.0
transformers==4.29.2
huggingface-hub==0.15.1

pycocotools==2.0.6
einops==0.3.0
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
clip==0.2.0
kornia==0.6.0

pydantic==1.10.5
test-tube==0.7.5